Add files via upload
This query calculates the infant mortality rate for each state in the year 2023. The query selects the state name and calculates the infant mortality rate by dividing the number of infant deaths by the total number of births for each state. The WHERE clause filters the results to only include data from the year 2023